refactor MDXImageWrapper 컴포넌트 organisms/MDX 디렉터리로 이동
