+++ src/components/organisms/MDX/MDXImageWrapper/MDXImageWrapper.style.ts
@@ -1,18 +1,5 @@
-import { PropsWithChildren } from 'react';
-import styled from '@emotion/styled';
 import { css } from '@emotion/react';
-import { MDXImageWrapperProps } from 'components/@shared/MDXImageWrapper/MDXImageWrapper.type';
-
-const MDXImageWrapper = ({ children, style, caption }: PropsWithChildren<MDXImageWrapperProps>) => {
-  return (
-    <Wrapper style={style}>
-      <ImageWrapper>{children}</ImageWrapper>
-      <ImageCaption>{caption}</ImageCaption>
-    </Wrapper>
-  );
-};
-
-export default MDXImageWrapper;
+import styled from '@emotion/styled';
 
 const Wrapper = styled.div`
   max-width: 450px;
@@ -47,3 +34,5 @@ const ImageCaption = styled.figcaption`
     text-align: center;
   `}
 `;
+
+export { Wrapper, ImageWrapper, ImageCaption };
