+++ src/components/organisms/MDX/MDXImageWrapper/MDXImageWrapper.type.ts
